+++ src/navigationBanner/Table.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import ItemDetailsView from '../components/itemDetailsView/itemDetailsView';
 import archived from '../images/archived.png'
 import { AccountContext } from '../util/Accounts';
 import './Table.css'
@@ -9,8 +10,9 @@ class Table extends Component {
         super(props)
         this.state = {
             Archived_Items: [], loading: false,
-            Headers: ['Name', 'Collection', 'Notes', 'Archived'],
+            Headers: ['Name', 'Collection', 'Notes', 'Archived'], item: [], editItem: false
         }
+        this.toggleDetailsView = this.toggleDetailsView.bind(this);
     }
 
     componentDidMount(){
@@ -26,6 +28,10 @@ class Table extends Component {
             });
     }
 
+    toggleDetailsView() {
+        this.setState({ editItem: !this.state.editItem });
+    }
+
     getItems = async (email) => {
         let queryURL = 'https://3cv3j619jg.execute-api.us-east-2.amazonaws.com/test/inventorme-items?userEmail="' + email + '"';
         const response = await fetch(queryURL.toString());
@@ -37,6 +43,10 @@ class Table extends Component {
         return archivedItems;
     }
 
+    filterItemByID(ID) {
+        this.setState({item: this.state.Archived_Items.filter(item => item.itemID === ID), editItem: true});
+    }
+
     renderTableHeader(){
         return this.state.Headers.map((key,index) => {
             return <th key={index}>{key.toUpperCase()}</th>
@@ -46,15 +56,18 @@ class Table extends Component {
     render() {
         return (
             <div>
+                { this.state.editItem ? 
+                <ItemDetailsView toggleDetailsView = {this.toggleDetailsView} editItem = {this.state.item} archive={true}/> :
+                null }
                 <h1 id = 'title'>Archived Items</h1>
                 { this.state.loading ?
                 <div className="loading-container"> <div className="form-load-symbol"/></div>
                 : null }
-                <table id= 'Archived_Items' style={{ marginBottom: '12em'}}>
+                <table id= 'Archived_Items' style={{ marginBottom: '12em', cursor: 'pointer'}}>
                 <tbody>
                 <tr>{this.renderTableHeader()}</tr>
                 { this.state.Archived_Items ? this.state.Archived_Items.map((Archived_Item) => (
-                <tr key = {Archived_Item.itemName}>
+                <tr key = {Archived_Item.itemName} onClick={() => this.filterItemByID(Archived_Item.itemID)}>
                 <td>{Archived_Item.itemName}</td> 
                 <td>{Archived_Item.itemCategory}</td>
                 <td>{Archived_Item.itemNotes}</td>
